最后一些优化，等后续新增接口

+++ model/dish.go
@@ -1,10 +1,10 @@
 package model
 
 type Dish struct {
-	DishNo    int    `json:"dish_no"`
-	DishName  string `json:"dish_name"`
-	DishPrice int    `json:"dish_price"`
-	DishTaste string `json:"dish_taste"`
+	DishNo    int     `json:"dish_no"`
+	DishName  string  `json:"dish_name"`
+	DishPrice float64 `json:"dish_price"`
+	DishTaste string  `json:"dish_taste"`
 }
 
 type PostDish struct {
